refactor(forks): Move ForkRangeDescriptor, get_fork_by_name
